rename to arena_rank
