+++ tests/test_hello.py
@@ -1,4 +1,4 @@
-from arena.hello import hello
+from arena_rank.hello import hello
 
 def test_hello() -> None:
     assert hello() == "Hello from the arena"
